impl. new way to solve the problem, creating hook
The Playlist has "one-to-many" relationship with the document  of Group,
one playlist can only be assigned to one group and one group can have a
many playlists

We need:
On the main page, we need to display a group of
Recent playlists and a group of Popular playlists.

We connected playlist and group documents through the field "groupID"
in the Playlist document, so currently it shows playlists of the groups
thanks to filtering using groupID, It also means that if one playlist
is in Latest groups, it can't be in the Popular group, which is not very
pleasant.

But we still required to make Recent and Popular groups and Group page
which relies on the 'groupID' field should not be affected

Solution:
I made a hook, which fetches Latest playlists and used it instead of
groupID for Group component and the Group page.

If groupID is not defined, then Group component's link will refer to the
`group/${title}` instead of `group/${groupID}` path, where title is the
title of the group, then on the Group page itself we use this to check
if path's id `group/:id` is equals to "Latest", if yes then we use hook
if not then default behaviour goes to play